Base Bar - Mana, Health & Chatbox
Finished the Mana and Health container of the basebar, they now display
properly and do not extend outside their boarders. Also added the
chatbox and must work some more on adding some sort of design for it.
Mana and Health text needs a bit of work. Code has also had some clean
ups, and will no longer use showNotification() and will require every
instance of that to be moved towards the ChatBox to give it the
Runescape Feel.