Working on Inventory and Fixing Name Bug in iOS Version
As previously stated in an issue, there is a problem with the username
located above the user when creating a new user. This problem only
seems to occur for levels 1-8 and through whatever reason doesn’t occur
afterwards. The update above is yet to be completed (have to fix
clicking on upscale of 1). This is being fixed through the creation of
top text and text canvas’ that are above the typical background,
foreground and entities canvas. Looking at the code should explain it
further.