Enchantment & Shopping
This update consists of a continuance in my adventure of refactoring
the client. Both the store and enchantment interfaces are bug free and
fixed in accordance to all the bugs reported by the players.
Furthermore, the way items are passed onto the game server has been
revamped to better suit the new format. The inventory frame and other
classes have been converted to fully utilize the object-oriented
paradigm. Less files have been removed until further notice.